fix: set function parameters for testing purposes too Fixes #8, #9, #10
